refactor: Put FOps into TimeSyntax and subpackage syntax
