+++ cats-effect/src/main/scala/com/avast/sst/catseffect/syntax/TimeSyntax.scala
@@ -1,4 +1,4 @@
-package com.avast.sst
+package com.avast.sst.catseffect.syntax
 
 import java.util.concurrent.TimeUnit
 
@@ -6,13 +6,19 @@ import cats.effect.syntax.bracket._
 import cats.effect.{Bracket, Clock}
 import cats.syntax.flatMap._
 import cats.syntax.functor._
+import com.avast.sst.catseffect.syntax.TimeSyntax.FOps
 
 import scala.concurrent.duration.Duration
 
-package object catseffect {
+trait TimeSyntax {
 
-  /**  */
-  implicit class FOps[F[_], A](private val f: F[A]) extends AnyVal {
+  implicit def sstFOps[F[_], A](f: F[A]): FOps[F, A] = new FOps(f)
+
+}
+
+object TimeSyntax {
+
+  final class FOps[F[_], A](private val f: F[A]) extends AnyVal {
 
     /** Measures the time it takes the effect to finish and records it using the provided function. */
     def time(record: Duration => F[Unit])(implicit F: Bracket[F, Throwable], C: Clock[F]): F[A] = {
