+++ cats-effect/src/main/scala/com/avast/sst/catseffect/syntax/package.scala
@@ -1,0 +1,7 @@
+package com.avast.sst.catseffect
+
+package object syntax {
+
+  object time extends TimeSyntax
+
+}
